[TASK] Update to Apache Solr 9.8.0
Adjusts the following tests not using the provided solrconfig testing
file:
- AccessFilterQParserMultiValueTest
- AccessFilterQParserSingleValueTest
- AccessFilterQParserWithRootLineTest

As the lock type is not set if the solrconfig file is not used, the
directoryFactory RAMDirectoryFactory used as default by SolrTestCaseJ4
will fail as it cannot be used with the default lock type.

As the RAMDirectory is deprecated since Solr 7.6.0 we switch to the
MockDirectoryFactory which is intended to be used with tests.

This change is still compatible with Apache Solr 9.6.